replace original `getDefaultFormStateTest` test with refactored
