+++ packages/utils/test/schema/index.ts
@@ -1,5 +1,4 @@
 import getDefaultFormStateTest from './getDefaultFormStateTest';
-import getDefaultFormStateTest2 from './getDefaultFormStateTest2';
 import getDisplayLabelTest from './getDisplayLabelTest';
 import getClosestMatchingOptionTest from './getClosestMatchingOptionTest';
 import getFirstMatchingOptionTest from './getFirstMatchingOptionTest';
@@ -16,7 +15,6 @@ export * from './types';
 
 export {
   getDefaultFormStateTest,
-  getDefaultFormStateTest2,
   getDisplayLabelTest,
   getClosestMatchingOptionTest,
   getFirstMatchingOptionTest,
